// DBS-Gateway-BE-001/src/interfaces/application/applicationByCompanyAdmin.interface.ts
export interface IPersonalDetails {
  // CompanyAdminID:string;
  // ApplicantId:string;
  // AssignTo:string;
  // AssignBy:string;
  // CreateBy:string;
  // CounterSignatory:string;
  title: string;
  firstName: string;
  IstMidName: string;
  secMidName: string;
  thirdMidName: string;
  presentSurname: string;
  surNameAtBirth: string;
  usedUntil: string;
  forenamesIs: string;
  otherSurnameIs: string;
  driverLicenceIs: string;
  validPassportIs: string;
  gender: string;
  dateOfBirth: string;
  isAgreedPrivacyPolicy: boolean;
  isAgreedProtectionPolicy: boolean;
  isAgreedTermsConditions: boolean;
}

export interface IAddressDetails {
  CurrentDetails: {
    country: string;
    postCode: string;
    addressLine1: string;
    addressLine2: string;
    cityOrTown: string;
    residentFrom: string;
  };
  PreviousAddress: [
    {
      unusualAddress: string;
      country: string;
      postCode: string;
      addressLine1: string;
      addressLine2: string;
      cityOrTown: string;
      residentFrom: string;
      residentUntil: string;
    },
  ];
}
export interface IAdditionalDetails {
  birthCountry: string;
  birthCounty: string;
  birthTown: string;
  birthNationality: string;
  ukInsuranceNum: string;
  telephoneNum: string;
  mobileNumber: string;
  isUkNational: string;
  isPaidOrVoluntaryWork: string;
  isGuideNeeded: string;
}
export interface IIdentityCheck {
  passportNum: string;
  birthOnPassport: string;
  passportNationality: string;
  passportIssueDate: string;
  licenceNumber: string;
  dateOfBirthOnLicence: string;
  licenceType: string;
  licenceValidFrom: string;
  licenceIssueCountry: string;
}
export interface IisNationalUK {
  isPassportValid: boolean;
  isBiometricResidencePermit: boolean;
  isCurrentDrivingLicense1: boolean;
  idBirthCertificateIssued: boolean;
  idAdoptionCertificate: boolean;
  isCurrentDrivingLicense: boolean;
  isBiometricCurrentDrivingLicense: boolean;
  isBirthCertificateIssued: boolean;
  isPertnershipCertificate: boolean;
  isImmigrationDocuments: boolean;
  IsMortgageStatment: boolean;
  isIssuedBankBuildingSocity: boolean;
  isIssuedBankBuildingSocityForOutSideUK: boolean;
  isIssuedBankAccountOpeningConfirmation: boolean;
  isCreditCardStatment: boolean;
  isP45OrP60Statement: boolean;
  isCouncitTaxStatement: boolean;
  isLetterOfSponsorship: boolean;
  isUtiliyBill: boolean;
  isBenifitStatementForChild: boolean;
  isLocalDocGiven: boolean;
  isEEANationalIDCard: boolean;
  isCardCarringPassLogo: boolean;
  letterFromHeadTeacher: boolean;
  immigrationDocumentOrWorkPermit: boolean;
  merriageCivilPertnershipCertificate: boolean;
  immigrationDocOrWorkPermitFor16YrOlds: boolean;
}